replace logic
